+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - toolbar to new and edit file pages
 - PyMdown SaneHeaders Extension
 - repository type on the Home page
-- показувати тип репозиторію за допомогою бейджів
+- PrivatePublish model
+
 ### Changed
 
 ### Deprecated